Ignore screenkey-git
Some days ago, I think, screenkey broke due to some problems with it's
modules or something. The fix was installing 'screenkey' from the Arch
packages instead of 'screenkey-git' from the Arco packages. Some tests
(though not many) since then have not shown screenkey-git to work again,
but pacman wants to install it instead of screenkey, thus it will be
ignored for normal updates, because the extra 'N' is annoying.