RFC Viewer
Now this is quite nice to have. There is a `rfc` package in the normal
Arch repos, though that contains ~600 MiB of pdf and txt files and no
script or program to display them. So to emulate a man-like viewer a
number is provided, that is curl-ed from rfc-editor.org, written to a
temp file and then displayed with less. It may be feasible to just
directyl pipe to less, but we'll see.