sxhkd: switch xkbmap switcher to Menu-key
funny story: I tried this with xev and misclicked and got Super_R
instead. I then switched super to Super_L and this here to Super_R, but
Super_L seems to be undefined and thus all super hotkeys dropped the
supper, i.e. open the terminal with Return, close a window with q.

Had to use thunar to drop the repo-version into my .config to fix it :)